+++ app/src/main/java/com/ala158/magicpantry/data/ShoppingListItem.kt
@@ -20,7 +20,7 @@ data class ShoppingListItem(
     var shoppingListItemId: Long = 0L,
 
     @ColumnInfo(name = "item_amount")
-    var itemAmount: Int = 0,
+    var itemAmount: Double = 0.0,
 
     @ColumnInfo(name = "is_item_bought")
     var isItemBought: Boolean = false,
@@ -28,7 +28,7 @@ data class ShoppingListItem(
     @ColumnInfo(name = "related_ingredient_id", index = true)
     var relatedIngredientId: Long = 0L,
 ) {
-    constructor(amount: Int, isItemBought: Boolean) : this() {
+    constructor(amount: Double, isItemBought: Boolean) : this() {
         this.itemAmount = amount
         this.isItemBought = isItemBought
     }
