+++ app/src/main/java/com/ala158/magicpantry/ui/manualingredientinput/edit/ManualIngredientInputEditFragment.kt
@@ -77,7 +77,7 @@ class ManualIngredientInputEditFragment : Fragment() {
 
     private fun saveIngredient() {
         val name = nameEditText.editText?.text.toString()
-        val amount = amountEditText.editText?.text.toString().toInt()
+        val amount = amountEditText.editText?.text.toString().toDouble()
         val unit = unitEditDropdown.text.toString()
         val price = priceEditText.editText?.text.toString().toDouble()
