+++ app/src/main/java/com/ala158/magicpantry/ui/manualingredientinput/edit/PantryEditIngredientActivity.kt
@@ -186,9 +186,9 @@ class PantryEditIngredientActivity : AppCompatActivity() {
         textInputEditAmount.addTextChangedListener(object : TextWatcher {
             override fun afterTextChanged(s: Editable?) {
                 val amountString = s.toString()
-                var amount = 0
+                var amount = 0.0
                 if (amountString != "")
-                    amount = amountString.toInt()
+                    amount = amountString.toDouble()
                 pantryEditIngredientViewModel.ingredientEntry.value!!.setAmount(amount)
                 amountLabel.setTextColor(resources.getColor(R.color.mp_textview_grey, null))
                 isAmountValid = true
